Optimize article writer for website
